🌱 relay: extract pure pane classifier into bin/lib/pane-classifier.js
Adds bin/contributor-relay.sh back as a thin POSIX-sh compat wrapper
(`exec node .../contributor-relay.js "$@"`) so every deployment,
container image, and install step still invoking the old path keeps
working unchanged (kubestellar/hive#6429).

Extracts the pure pane-classification layer out of the relay into a
new bin/lib/pane-classifier.js CommonJS module with no process/tmux
access: the PANE_STATE_*/BLOCKED_REASON_* constants, paneTail, the
per-backend readiness/login/onboarding tables, blockingPromptKey,
paneShowsTransientAPIError/paneShowsUnretryableAPIError/
paneShowsLoginRequiredError and TRANSIENT_API_ERROR_TAIL_LINES,
classifyBlockedOnHumanReason/paneLooksBlockedOnHuman, and
classifyTmuxPane (now classifyPane(text, backend, deps)). The relay
require()s the module and keeps thin wrappers at the original call
sites (blockingPromptKey, getCLIState, classifyTmuxPane) that supply
the backend and the impure dependencies the pure classifier cannot
have: getCLIState still owns capturePaneText() before handing text to
the module's classifyReadiness(); classifyTmuxPane's bob branch used to
call bobIsRunning() (process-table access) directly, so classifyPane
now takes it as an injected deps.bobIsRunning(). This is a move, not a
rewrite — behaviour is unchanged.

HEADLESS_BACKENDS (one-shot CLI dispatch) and the WS protocol/auth and
tmux-session management (tmuxSessionHumanPresence,
tmuxSessionHasAttachedClient, tmuxSendNudge, etc.) are out of scope for
pane classification and stay in the relay; the latter is a reasonable
follow-up but not mechanical enough to be "cheap" here.

Updates every reference to the old bin/contributor-relay.sh path that
is about the program itself (Justfile, bin/contributor-agent.sh,
bin/test_backend_smoke.sh, docs, and the Go tests/comments that read
the relay's real source or reference its path) to contributor-relay.js.
References that must keep working against the compat wrapper (the
wrapper's own header, bin/README.md's description of it, and
src/Dockerfile.contributor, which now ships contributor-relay.js,
contributor-relay.sh, and bin/lib/ together since
contributor-agent.sh's SCRIPT_DIR resolves inside the container) are
left pointing at the .sh path deliberately. CHANGELOG.md and
src/docs/podman-host-execution-gap.md's literal historical ps-output
quote are untouched, per convention.

Adds a dedicated test section to bin/contributor-relay.test.js that
require()s bin/lib/pane-classifier.js directly and exercises
classifyReadiness, blockingPromptKey, classifyPane (including the
injected bobIsRunning dependency), paneTail,
classifyBlockedOnHumanReason, and the API-error detectors with no
relay/tmux/ws stubbing. The much larger set of existing
loadRelay()-based tests that intentionally interleave classification
assertions with wiring behaviour (nudge dispatch, __crashTick, restart
paths, etc.) is left as-is — splitting those apart is a separable,
low-risk follow-up, not something this PR's scope requires.

Verified: node bin/contributor-relay.test.js (328/328), node --check on
every touched .js file, bash -n and shellcheck on the compat wrapper,
go build ./..., go test ./pkg/agent/ -run Pane, go test
./pkg/dashboard/... ./pkg/logscrub/... ./pkg/github/..., go vet on the
touched packages, python3 src/scripts/check-docs-links.py on both
src/docs and docs, and just --list.

Fixes #6429
Refs #6427
