+++ changelog.d/changed-6429-relay-js-modules.md
@@ -1,0 +1,1 @@
+- `bin/contributor-relay.sh` — 4,269 lines of Node.js despite the `.sh` name, with no module boundaries — is now `bin/contributor-relay.js` ([#6429](https://github.com/hivecommons/hive/issues/6429)). `git mv` carries its history; a thin POSIX-sh compat wrapper stays at the old `bin/contributor-relay.sh` path (`exec node .../contributor-relay.js "$@"`) because the path is baked into existing deployments, so nothing that still invokes the `.sh` path breaks. The pure pane-classification layer — `paneTail`, the per-backend readiness/login/onboarding tables, `blockingPromptKey`, the transient/unretryable/login-required API-error detectors, and `getCLIState`'s classification logic (now `classifyReadiness`/`classifyPane`) — moves into a new, side-effect-free `bin/lib/pane-classifier.js` CommonJS module with no `process`/tmux access, which the relay now `require`s; behaviour is unchanged, only import paths and internal wiring moved. WS protocol/auth and tmux-session management are left in the relay for a follow-up.
