+++ config/backends.conf
@@ -211,8 +211,8 @@ claude_family_local_perm_flag_shell() {
 # sandbox (MXC: Seatbelt on macOS, bubblewrap on Linux — same underlying
 # bubblewrap dependency the Claude local sandbox above already requires),
 # turned on with the `--sandbox` flag introduced in copilot-cli 1.0.60 and
-# usable in the headless `-p` mode contributor-relay.sh already launches
-# Copilot with (`copilot: { flag: '-p' }` in bin/contributor-relay.sh).
+# usable in the headless `-p` mode contributor-relay.js already launches
+# Copilot with (`copilot: { flag: '-p' }` in bin/contributor-relay.js).
 # `--no-sandbox` is also a real flag, so a stale/older `copilot` binary
 # silently ignoring `--sandbox` is not something to assume — this function
 # verifies the installed CLI actually understands the flag before relying on
